+++ charts/api-gateway/values.yaml
@@ -1,5 +1,5 @@
 Namespace: api-gateway
 etcdDns: etcd-headless.core.svc.cluster.local
 tracingEndpoint: collector.linkerd-jaeger:55678
-branchNameTag: "main"
+branchNameTag: "pythonlib"
 dockerArtifactAccount: "dynamos1"
