+++ charts/orchestrator/values.yaml
@@ -2,5 +2,5 @@ Namespace: orchestrator
 hostPath: default
 etcdDns: etcd-headless.core.svc.cluster.local
 tracingEndpoint: collector.linkerd-jaeger:55678
-branchNameTag: "main"
+branchNameTag: "pythonlib"
 dockerArtifactAccount: "dynamos1"
