+++ go/cmd/sql-aggregate/main.go
@@ -4,23 +4,31 @@ import (
 	"context"
 	"os"
 	"strconv"
-	"sync"
 
 	"github.com/Jorrit05/DYNAMOS/pkg/lib"
 	"github.com/Jorrit05/DYNAMOS/pkg/msinit"
 	pb "github.com/Jorrit05/DYNAMOS/pkg/proto"
-	"go.opencensus.io/trace"
-	"go.opencensus.io/trace/propagation"
 )
 
 var (
 	logger               = lib.InitLogger(logLevel)
-	config               *msinit.Configuration
 	COORDINATOR          = make(chan struct{})
 	NR_OF_DATA_PROVIDERS = getNrOfDataProviders()
-	receiveMutex         = &sync.Mutex{}
-	mscommList           = []*pb.MicroserviceCommunication{}
+	msCommList           = []*pb.MicroserviceCommunication{}
 )
+
+func getNrOfDataProviders() int {
+	nr_of_data_providers_int := 0
+	nr_of_data_providers := os.Getenv("NR_OF_DATA_PROVIDERS")
+	var err error
+	if nr_of_data_providers != "" {
+		nr_of_data_providers_int, err = strconv.Atoi(nr_of_data_providers)
+		if err != nil {
+			logger.Sugar().Errorf("Error converting nr_of_data_providers to int: %v", err)
+		}
+	}
+	return nr_of_data_providers_int
+}
 
 func main() {
 	logger.Sugar().Debugf("Starting %s service", serviceName)
@@ -53,15 +61,31 @@ func messageHandler(config *msinit.Configuration) func(ctx context.Context, msCo
 		// Wait till all services and connections have started
 		<-COORDINATOR
 
-		switch msComm.RequestType {
-		case "sqlDataRequest":
-			err := handleSqlDataRequest(ctx, msComm)
-			if err != nil {
-				logger.Sugar().Errorf("Failed to process %s message: %v", msComm.RequestType, err)
-			}
+		msCommList = append(msCommList, msComm)
+		logger.Sugar().Infof("msCommList: %v", msCommList)
+		logger.Sugar().Infof("amount of data providers %v", NR_OF_DATA_PROVIDERS)
+		logger.Sugar().Infof("Length of msCommList %v", len(msCommList))
 
-		default:
-			logger.Sugar().Errorf("Unknown RequestType type: %v", msComm.RequestType)
+		// If NR_OF_DATA_PROVIDERS == 0 aggregate won't actually function and pass on the message.
+		// This can happen at this moment if the aggregate flag is set to True, but it is not allowed by policy.
+		if len(msCommList) == NR_OF_DATA_PROVIDERS || NR_OF_DATA_PROVIDERS == 0 {
+			logger.Sugar().Debugf(msCommList[0].Data.String())
+			// All messages have arrived
+			logger.Sugar().Infof("All messages have arrived, %v", len(msCommList))
+
+			switch msComm.RequestType {
+			case "sqlDataRequest":
+				ctx, msComm, err = handleSqlDataRequest(ctx, msCommList)
+				if err != nil {
+					logger.Sugar().Errorf("Failed to process %s message: %v", msComm.RequestType, err)
+				}
+
+			default:
+				logger.Sugar().Errorf("Unknown RequestType type: %v", msComm.RequestType)
+			}
+		} else {
+			logger.Sugar().Infof("Waiting for %v messages", NR_OF_DATA_PROVIDERS-len(msCommList))
+			return nil
 		}
 
 		config.NextClient.SendData(ctx, msComm)
@@ -69,34 +93,4 @@ func messageHandler(config *msinit.Configuration) func(ctx context.Context, msCo
 		close(config.StopMicroservice)
 		return nil
 	}
-}
-
-func getNrOfDataProviders() int {
-	nr_of_data_providers_int := 0
-	nr_of_data_providers := os.Getenv("NR_OF_DATA_PROVIDERS")
-	var err error
-	if nr_of_data_providers != "" {
-		nr_of_data_providers_int, err = strconv.Atoi(nr_of_data_providers)
-		if err != nil {
-			logger.Sugar().Errorf("Error converting nr_of_data_providers to int: %v", err)
-		}
-	}
-	return nr_of_data_providers_int
-}
-
-// This is the function being called by the last microservice
-func handleSqlDataRequest(ctx context.Context, msCommList []*pb.MicroserviceCommunication) (context.Context, *pb.MicroserviceCommunication, error) {
-	ctx, span := trace.StartSpan(ctx, "aggregtate/handleSqlDataRequest")
-	defer span.End()
-	logger.Sugar().Infof("Start %s handleSqlDataRequest", serviceName)
-
-	if len(msCommList) < 2 {
-		return ctx, msCommList[0], nil
-	}
-
-	// Coordinator ensures all services are started before further processing messages
-	msCommList[0].Traces["binaryTrace"] = propagation.Binary(span.SpanContext())
-	mergedMsComm := mergeData(msCommList)
-
-	return ctx, mergedMsComm, nil
 }
